requirements.txt
jax
pgx
mctx
optax
flax
hydra-core
chex
chess
numpy
omegaconf
tqdm
matplotlib
orbax-checkpoint
flashbax
pygame